Fix accidental cyclic dependency between object and trait
